# api/openapi.yaml
openapi: 3.0.3
info:
  title: Field Manager API
  description: 圃場管理システムAPI
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: /
    description: 相対パス

tags:
  - name: health
    description: ヘルスチェック
  - name: fields
    description: 圃場管理
  - name: imports
    description: wagriデータインポート
  - name: clusters
    description: H3クラスタリング

# セキュリティ定義(認証なしを明示)
security: []

paths:
  /health:
    get:
      tags:
        - health
      summary: ヘルスチェック
      operationId: healthCheck
      security: []
      responses:
        "200":
          description: サーバーが正常に動作している
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "429":
          description: リクエスト過多
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: サービス利用不可
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/fields:
    get:
      tags:
        - fields
      summary: 圃場一覧取得
      operationId: listFields
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: 圃場一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldListResponse"
        "400":
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/fields/{fieldId}:
    get:
      tags:
        - fields
      summary: 圃場詳細取得
      operationId: getField
      security: []
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 圃場詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
        "404":
          description: 圃場が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/imports:
    post:
      tags:
        - imports
      summary: インポートリクエスト
      description: wagri APIからのデータインポートをリクエストする
      operationId: requestImport
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "202":
          description: インポートリクエスト受付
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
        "400":
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/imports/{importId}:
    get:
      tags:
        - imports
      summary: インポートステータス取得
      description: インポートジョブのステータスを取得する
      operationId: getImportStatus
      security: []
      parameters:
        - name: importId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: インポートステータス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportStatus"
        "404":
          description: インポートジョブが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/clusters:
    get:
      tags:
        - clusters
      summary: クラスター一覧取得
      description: |
        指定されたズームレベルとバウンディングボックス内のH3クラスターを取得する。
        ズームレベルに応じて最適なH3解像度を自動選択する。
      operationId: getClusters
      security: []
      parameters:
        - name: zoom
          in: query
          required: true
          description: Google Mapsのズームレベル(1.0-22.0、少数対応)
          schema:
            type: number
            format: double
            minimum: 1.0
            maximum: 22.0
        - name: sw_lat
          in: query
          required: true
          description: 南西端の緯度
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: sw_lng
          in: query
          required: true
          description: 南西端の経度
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: ne_lat
          in: query
          required: true
          description: 北東端の緯度
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: ne_lng
          in: query
          required: true
          description: 北東端の経度
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
      responses:
        "200":
          description: クラスター一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterListResponse"
        "400":
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    Field:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
        areaHa:
          type: number
          format: double
          description: 面積(ヘクタール)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    FieldListResponse:
      type: object
      required:
        - fields
        - total
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/Field"
        total:
          type: integer

    ImportRequest:
      type: object
      required:
        - cityCode
      properties:
        cityCode:
          type: string
          description: 市区町村コード
          example: "163210"

    ImportResponse:
      type: object
      required:
        - importId
      properties:
        importId:
          type: string
          format: uuid
          description: インポートジョブID
        executionArn:
          type: string
          description: Step Functions実行ARN

    ImportStatus:
      type: object
      required:
        - id
        - cityCode
        - status
        - processedRecords
        - failedRecords
        - progress
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        cityCode:
          type: string
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
            - partially_completed
        totalRecords:
          type: integer
          nullable: true
        processedRecords:
          type: integer
        failedRecords:
          type: integer
        progress:
          type: number
          format: double
          description: 進捗率(0-100)
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true

    Cluster:
      type: object
      required:
        - h3Index
        - lat
        - lng
        - count
      properties:
        h3Index:
          type: string
          description: H3インデックス(16進数文字列)
          example: "831f8dfffffffff"
        lat:
          type: number
          format: double
          description: クラスター中心の緯度
        lng:
          type: number
          format: double
          description: クラスター中心の経度
        count:
          type: integer
          description: クラスターに含まれる圃場数

    ClusterListResponse:
      type: object
      required:
        - clusters
        - isStale
      properties:
        clusters:
          type: array
          items:
            $ref: "#/components/schemas/Cluster"
        isStale:
          type: boolean
          description: クラスターが再計算中かどうか