# internal/generated/openapi/server.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package openapi

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// クラスター一覧取得
	// (GET /api/v1/clusters)
	GetClusters(c *gin.Context, params GetClustersParams)
	// 圃場一覧取得
	// (GET /api/v1/fields)
	ListFields(c *gin.Context, params ListFieldsParams)
	// 圃場詳細取得
	// (GET /api/v1/fields/{fieldId})
	GetField(c *gin.Context, fieldId openapi_types.UUID)
	// インポートリクエスト
	// (POST /api/v1/imports)
	RequestImport(c *gin.Context)
	// インポートステータス取得
	// (GET /api/v1/imports/{importId})
	GetImportStatus(c *gin.Context, importId openapi_types.UUID)
	// ヘルスチェック
	// (GET /health)
	HealthCheck(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetClusters operation middleware
func (siw *ServerInterfaceWrapper) GetClusters(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetClustersParams

	// ------------- Required query parameter "zoom" -------------

	if paramValue := c.Query("zoom"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument zoom is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "zoom", c.Request.URL.Query(), &params.Zoom)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter zoom: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "sw_lat" -------------

	if paramValue := c.Query("sw_lat"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument sw_lat is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "sw_lat", c.Request.URL.Query(), &params.SwLat)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sw_lat: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "sw_lng" -------------

	if paramValue := c.Query("sw_lng"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument sw_lng is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "sw_lng", c.Request.URL.Query(), &params.SwLng)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sw_lng: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "ne_lat" -------------

	if paramValue := c.Query("ne_lat"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument ne_lat is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "ne_lat", c.Request.URL.Query(), &params.NeLat)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter ne_lat: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "ne_lng" -------------

	if paramValue := c.Query("ne_lng"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument ne_lng is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "ne_lng", c.Request.URL.Query(), &params.NeLng)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter ne_lng: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetClusters(c, params)
}

// ListFields operation middleware
func (siw *ServerInterfaceWrapper) ListFields(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListFieldsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", c.Request.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListFields(c, params)
}

// GetField operation middleware
func (siw *ServerInterfaceWrapper) GetField(c *gin.Context) {

	var err error

	// ------------- Path parameter "fieldId" -------------
	var fieldId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "fieldId", c.Param("fieldId"), &fieldId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter fieldId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetField(c, fieldId)
}

// RequestImport operation middleware
func (siw *ServerInterfaceWrapper) RequestImport(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.RequestImport(c)
}

// GetImportStatus operation middleware
func (siw *ServerInterfaceWrapper) GetImportStatus(c *gin.Context) {

	var err error

	// ------------- Path parameter "importId" -------------
	var importId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "importId", c.Param("importId"), &importId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter importId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetImportStatus(c, importId)
}

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheck(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/v1/clusters", wrapper.GetClusters)
	router.GET(options.BaseURL+"/api/v1/fields", wrapper.ListFields)
	router.GET(options.BaseURL+"/api/v1/fields/:fieldId", wrapper.GetField)
	router.POST(options.BaseURL+"/api/v1/imports", wrapper.RequestImport)
	router.GET(options.BaseURL+"/api/v1/imports/:importId", wrapper.GetImportStatus)
	router.GET(options.BaseURL+"/health", wrapper.HealthCheck)
}

type GetClustersRequestObject struct {
	Params GetClustersParams
}

type GetClustersResponseObject interface {
	VisitGetClustersResponse(w http.ResponseWriter) error
}

type GetClusters200JSONResponse ClusterListResponse

func (response GetClusters200JSONResponse) VisitGetClustersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetClusters400JSONResponse ErrorResponse

func (response GetClusters400JSONResponse) VisitGetClustersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetClusters500JSONResponse ErrorResponse

func (response GetClusters500JSONResponse) VisitGetClustersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type ListFieldsRequestObject struct {
	Params ListFieldsParams
}

type ListFieldsResponseObject interface {
	VisitListFieldsResponse(w http.ResponseWriter) error
}

type ListFields200JSONResponse FieldListResponse

func (response ListFields200JSONResponse) VisitListFieldsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListFields400JSONResponse ErrorResponse

func (response ListFields400JSONResponse) VisitListFieldsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type ListFields500JSONResponse ErrorResponse

func (response ListFields500JSONResponse) VisitListFieldsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetFieldRequestObject struct {
	FieldId openapi_types.UUID `json:"fieldId"`
}

type GetFieldResponseObject interface {
	VisitGetFieldResponse(w http.ResponseWriter) error
}

type GetField200JSONResponse Field

func (response GetField200JSONResponse) VisitGetFieldResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetField404JSONResponse ErrorResponse

func (response GetField404JSONResponse) VisitGetFieldResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetField500JSONResponse ErrorResponse

func (response GetField500JSONResponse) VisitGetFieldResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type RequestImportRequestObject struct {
	Body *RequestImportJSONRequestBody
}

type RequestImportResponseObject interface {
	VisitRequestImportResponse(w http.ResponseWriter) error
}

type RequestImport202JSONResponse ImportResponse

func (response RequestImport202JSONResponse) VisitRequestImportResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(202)

	return json.NewEncoder(w).Encode(response)
}

type RequestImport400JSONResponse ErrorResponse

func (response RequestImport400JSONResponse) VisitRequestImportResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type RequestImport500JSONResponse ErrorResponse

func (response RequestImport500JSONResponse) VisitRequestImportResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetImportStatusRequestObject struct {
	ImportId openapi_types.UUID `json:"importId"`
}

type GetImportStatusResponseObject interface {
	VisitGetImportStatusResponse(w http.ResponseWriter) error
}

type GetImportStatus200JSONResponse ImportStatus

func (response GetImportStatus200JSONResponse) VisitGetImportStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetImportStatus404JSONResponse ErrorResponse

func (response GetImportStatus404JSONResponse) VisitGetImportStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetImportStatus500JSONResponse ErrorResponse

func (response GetImportStatus500JSONResponse) VisitGetImportStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type HealthCheckRequestObject struct {
}

type HealthCheckResponseObject interface {
	VisitHealthCheckResponse(w http.ResponseWriter) error
}

type HealthCheck200JSONResponse HealthResponse

func (response HealthCheck200JSONResponse) VisitHealthCheckResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type HealthCheck429JSONResponse ErrorResponse

func (response HealthCheck429JSONResponse) VisitHealthCheckResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(429)

	return json.NewEncoder(w).Encode(response)
}

type HealthCheck503JSONResponse ErrorResponse

func (response HealthCheck503JSONResponse) VisitHealthCheckResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// クラスター一覧取得
	// (GET /api/v1/clusters)
	GetClusters(ctx context.Context, request GetClustersRequestObject) (GetClustersResponseObject, error)
	// 圃場一覧取得
	// (GET /api/v1/fields)
	ListFields(ctx context.Context, request ListFieldsRequestObject) (ListFieldsResponseObject, error)
	// 圃場詳細取得
	// (GET /api/v1/fields/{fieldId})
	GetField(ctx context.Context, request GetFieldRequestObject) (GetFieldResponseObject, error)
	// インポートリクエスト
	// (POST /api/v1/imports)
	RequestImport(ctx context.Context, request RequestImportRequestObject) (RequestImportResponseObject, error)
	// インポートステータス取得
	// (GET /api/v1/imports/{importId})
	GetImportStatus(ctx context.Context, request GetImportStatusRequestObject) (GetImportStatusResponseObject, error)
	// ヘルスチェック
	// (GET /health)
	HealthCheck(ctx context.Context, request HealthCheckRequestObject) (HealthCheckResponseObject, error)
}

type StrictHandlerFunc = strictgin.StrictGinHandlerFunc
type StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetClusters operation middleware
func (sh *strictHandler) GetClusters(ctx *gin.Context, params GetClustersParams) {
	var request GetClustersRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetClusters(ctx, request.(GetClustersRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetClusters")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetClustersResponseObject); ok {
		if err := validResponse.VisitGetClustersResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListFields operation middleware
func (sh *strictHandler) ListFields(ctx *gin.Context, params ListFieldsParams) {
	var request ListFieldsRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.ListFields(ctx, request.(ListFieldsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListFields")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(ListFieldsResponseObject); ok {
		if err := validResponse.VisitListFieldsResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetField operation middleware
func (sh *strictHandler) GetField(ctx *gin.Context, fieldId openapi_types.UUID) {
	var request GetFieldRequestObject

	request.FieldId = fieldId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetField(ctx, request.(GetFieldRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetField")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetFieldResponseObject); ok {
		if err := validResponse.VisitGetFieldResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// RequestImport operation middleware
func (sh *strictHandler) RequestImport(ctx *gin.Context) {
	var request RequestImportRequestObject

	var body RequestImportJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RequestImport(ctx, request.(RequestImportRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RequestImport")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(RequestImportResponseObject); ok {
		if err := validResponse.VisitRequestImportResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetImportStatus operation middleware
func (sh *strictHandler) GetImportStatus(ctx *gin.Context, importId openapi_types.UUID) {
	var request GetImportStatusRequestObject

	request.ImportId = importId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetImportStatus(ctx, request.(GetImportStatusRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetImportStatus")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetImportStatusResponseObject); ok {
		if err := validResponse.VisitGetImportStatusResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// HealthCheck operation middleware
func (sh *strictHandler) HealthCheck(ctx *gin.Context) {
	var request HealthCheckRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.HealthCheck(ctx, request.(HealthCheckRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "HealthCheck")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(HealthCheckResponseObject); ok {
		if err := validResponse.VisitHealthCheckResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}