# internal/generated/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package openapi

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ImportStatusStatus.
const (
	Completed          ImportStatusStatus = "completed"
	Failed             ImportStatusStatus = "failed"
	PartiallyCompleted ImportStatusStatus = "partially_completed"
	Pending            ImportStatusStatus = "pending"
	Processing         ImportStatusStatus = "processing"
)

// Cluster defines model for Cluster.
type Cluster struct {
	// Count クラスターに含まれる圃場数
	Count int `json:"count"`

	// H3Index H3インデックス(16進数文字列)
	H3Index string `json:"h3Index"`

	// Lat クラスター中心の緯度
	Lat float64 `json:"lat"`

	// Lng クラスター中心の経度
	Lng float64 `json:"lng"`
}

// ClusterListResponse defines model for ClusterListResponse.
type ClusterListResponse struct {
	Clusters []Cluster `json:"clusters"`

	// IsStale クラスターが再計算中かどうか
	IsStale bool `json:"isStale"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// Field defines model for Field.
type Field struct {
	// AreaHa 面積(ヘクタール)
	AreaHa      *float64           `json:"areaHa,omitempty"`
	CreatedAt   time.Time          `json:"createdAt"`
	Description *string            `json:"description,omitempty"`
	Id          openapi_types.UUID `json:"id"`
	Name        string             `json:"name"`
	UpdatedAt   time.Time          `json:"updatedAt"`
}

// FieldListResponse defines model for FieldListResponse.
type FieldListResponse struct {
	Fields []Field `json:"fields"`
	Total  int     `json:"total"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// ImportRequest defines model for ImportRequest.
type ImportRequest struct {
	// CityCode 市区町村コード
	CityCode string `json:"cityCode"`
}

// ImportResponse defines model for ImportResponse.
type ImportResponse struct {
	// ExecutionArn Step Functions実行ARN
	ExecutionArn *string `json:"executionArn,omitempty"`

	// ImportId インポートジョブID
	ImportId openapi_types.UUID `json:"importId"`
}

// ImportStatus defines model for ImportStatus.
type ImportStatus struct {
	CityCode         string             `json:"cityCode"`
	CompletedAt      *time.Time         `json:"completedAt"`
	CreatedAt        time.Time          `json:"createdAt"`
	ErrorMessage     *string            `json:"errorMessage"`
	FailedRecords    int                `json:"failedRecords"`
	Id               openapi_types.UUID `json:"id"`
	ProcessedRecords int                `json:"processedRecords"`

	// Progress 進捗率(0-100)
	Progress     float64            `json:"progress"`
	StartedAt    *time.Time         `json:"startedAt"`
	Status       ImportStatusStatus `json:"status"`
	TotalRecords *int               `json:"totalRecords"`
}

// ImportStatusStatus defines model for ImportStatus.Status.
type ImportStatusStatus string

// GetClustersParams defines parameters for GetClusters.
type GetClustersParams struct {
	// Zoom Google Mapsのズームレベル(1.0-22.0、少数対応)
	Zoom float64 `form:"zoom" json:"zoom"`

	// SwLat 南西端の緯度
	SwLat float64 `form:"sw_lat" json:"sw_lat"`

	// SwLng 南西端の経度
	SwLng float64 `form:"sw_lng" json:"sw_lng"`

	// NeLat 北東端の緯度
	NeLat float64 `form:"ne_lat" json:"ne_lat"`

	// NeLng 北東端の経度
	NeLng float64 `form:"ne_lng" json:"ne_lng"`
}

// ListFieldsParams defines parameters for ListFields.
type ListFieldsParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// RequestImportJSONRequestBody defines body for RequestImport for application/json ContentType.
type RequestImportJSONRequestBody = ImportRequest